fix(csv): csv file display and build errors
